implement memo and cap instruction sets + fix cap leaks
Vm: i initially allocated captures with the vm.allocator.  however,
there were leaks which i wasn't able to easily fix.  so i added a
'captures_arena' param to exec() which gives ownership to the user,
allowing them to be deinit() together.

i believe this is a poor memory strategy likely returning a lot of
unused memory back along with for captures.  so i've left a TODO.

* memo: add Capture.ChildIterator
* pattern: add Memo() and Cap()
* create tests/ folder